teste_front/api/requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
azure-common==1.1.23
azure-core==1.5.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backcall==0.1.0
beautifulsoup4==4.9.3
boto3==1.15.15
botocore==1.18.15
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
cssutils==1.0.2
cycler==0.11.0
decorator==4.4.2
distlib==0.3.1
Django==3.0.5
django-admin-autocomplete-filter==0.6.1
django-admin-json-editor==0.2.0
django-admin-rangefilter==0.6.3
django-cacheops==4.2
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-cron==0.5.1
django-currentuser==0.5.1
django-deprecate-fields==0.1.1
django-dynamic-preferences==1.9
django-filter==2.2.0
django-health-check==3.12.1
django-hint==0.1.5
django-ipware==2.1.0
django-kiwi @ git+https://gitlab.com/luisccf/django-kiwi/@81bb3294ca6418527bcd9b06bba4a7a52912b3c4
django-lightweight-tests==0.0.1
django-mssql-backend==2.8.0
django-nested-admin==4.0.2
django-rest-passwordreset==1.1.0
django-simple-history==3.0.0
django-split-settings==1.0.0
django-sql-utils==0.6.1
django-sslserver==0.22
django-summernote==0.8.11.6
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-nested-routers==0.91
et_xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.0.2
filelock==3.0.12
firebase-admin==4.3.0
fonttools==4.33.3
freezegun==0.3.15
funcy==1.14
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
httplib2==0.18.1
idna==2.9
importlib-metadata==3.10.0
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.0
jmespath==0.10.0
jsonfield==3.1.0
kiwisolver==1.4.2
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
msgpack==1.0.0
msrest==0.6.13
newrelic==8.7.0
numpy==1.21.6
oauthlib==3.1.0
openpyxl==3.0.3
packaging==21.3
parameterized==0.7.4
parso==0.7.0
pbr==5.5.1
persisting-theory==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==9.1.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psycopg2-binary==2.9.9
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pydash==4.7.6
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyodbc==4.0.30
pyparsing==3.0.8
python-dateutil==2.8.1
python-monkey-business==1.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rollbar==0.15.0
rsa==4.6
s3transfer==0.3.3
sentry-sdk==1.8.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
stevedore==3.3.0
text-unidecode==1.3
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.1
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.22.0
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.4.1